Extract dumb polling to a custom hook
